fix warnings

+++ src/xtd_forms_native_wxwidgets/src/xtd/forms/native/wxwidgets/tab_control.cpp
@@ -33,12 +33,12 @@ void tab_control::insert_item(intptr_t control, size_t index, intptr_t page) {
 
 void tab_control::page_image_index(intptr_t control, size_t index, size_t image_index) {
   if (control == 0) return;
-  static_cast<wxNotebookBase*>(reinterpret_cast<control_handler*>(control)->control())->SetPageImage(index, image_index);
+  static_cast<wxNotebookBase*>(reinterpret_cast<control_handler*>(control)->control())->SetPageImage(static_cast<int32_t>(index), static_cast<int32_t>(image_index));
 }
 
 void tab_control::page_text(intptr_t control, size_t index, const std::string& text) {
   if (control == 0) return;
-  static_cast<wxNotebookBase*>(reinterpret_cast<control_handler*>(control)->control())->SetPageText(index, {text.c_str(), wxMBConvUTF8()});
+  static_cast<wxNotebookBase*>(reinterpret_cast<control_handler*>(control)->control())->SetPageText(static_cast<int32_t>(index), {text.c_str(), wxMBConvUTF8()});
 }
 
 size_t tab_control::selected_index(intptr_t control) {
